Hopefully a fix for damage-related errors.
AtherysCore no longer attempts to extend the Sponge damage-related interfaces and classes to do with DamageSource. Instead, only custom DamageTypes are provided. When a DamageEvent is detected, the DamageConfig is looked up by the ItemType of the item in the player's hand, and the appropriate DamageType is appended to the DamageEntityEvent's Cause.